Fix bug in ReplaceSquares func
ReplaceSquares function allowed to delete "non-existent" squares. It means
that DeletedCompletely function could be called to create vec with deleted
black squares and delete them again. It could lead to inefficiency of the
code.

Also, if-statement to check whether we clicked on the black square denies
prohibits replacing colourful square with deleted one.